+++ jitsi-meet/jitsi-meet.d/local/share/cook/bin/configure-prosody.sh
@@ -26,7 +26,8 @@ sep=$'\001'
 < "$TEMPLATEPATH/prosody.cfg.lua.in" \
   sed "s${sep}%%domain%%${sep}$DOMAIN${sep}g" | \
   sed "s${sep}%%email%%${sep}$EMAIL${sep}g" | \
-  sed "s${sep}%%keypassword%%${sep}$KEYPASSWORD${sep}g" \
+  sed "s${sep}%%keypassword%%${sep}$KEYPASSWORD${sep}g" | \
+  sed "s${sep}%%secpassword%%${sep}$SECPASSWORD${sep}g" \
   > /usr/local/etc/prosody/prosody.cfg.lua
 
 # setup prosody