+++ jitsi-meet/jitsi-meet.d/local/share/cook/templates/jicofostatus.sh
@@ -1,0 +1,3 @@
+#!/usr/local/bin/bash
+
+curl -s "http://localhost:8080/debug?full=true" | jq .